move the files to  folders
