+++ dbms/src/test/java/org/polypheny/db/cypher/DdlTest.java
@@ -34,6 +34,7 @@ import org.polypheny.db.TestHelper.JdbcConnection;
 import org.polypheny.db.catalog.Catalog;
 import org.polypheny.db.catalog.entity.logical.LogicalGraph;
 import org.polypheny.db.catalog.entity.logical.LogicalNamespace;
+import org.polypheny.db.cypher.clause.write.DmlInsertTest;
 import org.polypheny.db.webui.models.results.GraphResult;
 
 @Tag("adapter")
